marching cubes rbf - doubles
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@6000 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ surface/include/pcl/surface/marching_cubes_rbf.h
@@ -95,8 +95,8 @@ namespace pcl
 
     protected:
       /** \brief the Radial Basis Function kernel. */
-      float
-      kernel (Eigen::Vector3f c, Eigen::Vector3f x);
+      double
+      kernel (Eigen::Vector3d c, Eigen::Vector3d x);
 
       /** \brief The off-surface displacement value. */
       float off_surface_epsilon_;
